FROMLIST: dt-bindings: arm: qcom: Document HAMOA-IOT-EVK board
Document the device tree binding for the HAMOA-IOT-EVK board, which uses
the Qualcomm X1E80100 SoC.

The EVK consists of a carrier board and a modular System-on-Module (SoM).
The SoM integrates the SoC, PMICs, and essential GPIOs, while the EVK
carrier board provides additional peripherals such as UART and USB
interfaces.

Link: https://lore.kernel.org/all/20250729-hamoa_initial-v3-1-806e092789dc@oss.qualcomm.com
Signed-off-by: Yijie Yang <yijie.yang@oss.qualcomm.com>